fix: run evaluations with the same agent context as Studio
Conversation-evaluation runs invoked the agent with a legacy master
prompt and no tools, and extraction-evaluation runs duplicated (and had
drifted from) the Studio CSV extraction invocation, so evaluation
scores did not reflect real agent behaviour.

Extract StreamingService's LLM request building into a shared
AgentLlmRequestService (master prompt, tools, metadata, attachments)
with a runSingleTurn helper, and a shared
StructuredExtractionAgentRunnerService for CSV-row extraction, and use
them from both the Studio paths and the evaluation workers. The only
deliberate divergence left: session-metadata tools are excluded from
evaluation runs since there is no session row to mutate.
