+++ apps/api/src/domains/agents/csv-extraction-runs/agent-csv-extraction-run-processor.service.spec.ts
@@ -7,6 +7,7 @@ import {
   teardownE2eTestDatabase,
 } from "@/common/test/test-database"
 import { toAgentWithSettingsRunJobPayload } from "@/domains/agents/shared/agent-with-settings-run.helper"
+import { StructuredExtractionAgentRunnerService } from "@/domains/agents/shared/structured-extraction-agent-runner.service"
 import { documentFactory } from "@/domains/documents/document.factory"
 import { createOrganizationWithAgent } from "@/domains/organizations/organization.factory"
 import { LlmModule } from "@/external/llm/llm.module"
@@ -33,6 +34,7 @@ describe("AgentCsvExtractionRunProcessorService", () => {
       additionalImports: [LlmModule],
       providers: [
         AgentCsvExtractionRunProcessorService,
+        StructuredExtractionAgentRunnerService,
         { provide: AgentCsvExtractionRunStatusNotifierService, useValue: mockStatusNotifier },
         { provide: AgentCsvExtractionRunCsvExportService, useValue: mockCsvExport },
       ],
