+++ apps/api/src/domains/agents/shared/agent-session-messages/streaming/streaming.module.ts
@@ -9,28 +9,19 @@ import { ConversationAgentSession } from "@/domains/agents/conversation-agent-se
 import { ConversationAgentSessionsModule } from "@/domains/agents/conversation-agent-sessions/conversation-agent-sessions.module"
 import { FormAgentSession } from "@/domains/agents/form-agent-sessions/form-agent-session.entity"
 import { FormAgentSessionsModule } from "@/domains/agents/form-agent-sessions/form-agent-sessions.module"
-import { McpServersModule } from "@/domains/mcp-servers/mcp-servers.module"
-import { McpModule } from "@/external/mcp"
-import { AgentMessageAttachmentDocumentsService } from "../agent-message-attachment-documents.service"
+import { AgentLlmModule } from "./agent-llm.module"
 import { StreamingController } from "./streaming.controller"
 import { StreamingService } from "./streaming.service"
-import { ToolsService } from "./tools.service"
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([...moduleFeatures, ConversationAgentSession, FormAgentSession]),
     ...moduleImports,
     forwardRef(() => FormAgentSessionsModule),
-    McpModule,
-    McpServersModule,
+    forwardRef(() => AgentLlmModule),
     forwardRef(() => ConversationAgentSessionsModule),
   ],
-  providers: [
-    ...moduleProviders,
-    AgentMessageAttachmentDocumentsService,
-    StreamingService,
-    ToolsService,
-  ],
+  providers: [...moduleProviders, StreamingService],
   controllers: [StreamingController],
   exports: [StreamingService],
 })
